+++ src/utils/format.js
@@ -6,17 +6,6 @@ export const leftPad = (str, padString, targetLength) => {
   return str.toString().padStart(targetLength, padString);
 };
 
-export const createItems = (elementsData, getHtml) => {
-  const container = document.createDocumentFragment();
-  container.innerHTML = ``;
-
-  for (const elementData of elementsData) {
-    container.innerHTML += getHtml(elementData);
-  }
-
-  return container.innerHTML;
-};
-
 export const millisecondsToHm = (timeInMs) => {
   const hours = Math.floor(timeInMs / 3600000);
   const minutes = Math.floor(timeInMs % 3600000 / 60);
