+++ src/utils/render.js
@@ -41,3 +41,14 @@ export const replace = (newComponent, oldComponent) => {
     parentElement.replaceChild(newElement, oldElement);
   }
 };
+
+export const createItems = (elementsData, getHtml) => {
+    const container = document.createDocumentFragment();
+    container.innerHTML = ``;
+
+    for (const elementData of elementsData) {
+        container.innerHTML += getHtml(elementData);
+    }
+
+    return container.innerHTML;
+};
